# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the website of the FSV Informatik",
        "title": "Fachschaft Informatik API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/archive": {
            "get": {
                "tags": [
                    "Archive"
                ],
                "summary": "List archive entries",
                "operationId": "getArchive",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Program ID",
                        "name": "program_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Version",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Module ID",
                        "name": "module_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ArchiveEntryResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Archive"
                ],
                "summary": "Upload archive entry",
                "operationId": "postArchive",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CSRF Token",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Exam PDF",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exam Date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Module ID",
                        "name": "module_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "PO Version",
                        "name": "version",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Optional specific group ID to add revision to",
                        "name": "group_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Optional comment",
                        "name": "comment",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.ArchiveEntryResponse"
                        }
                    }
                }
            }
        },
        "/archive/files/{fileId}": {
            "delete": {
                "tags": [
                    "Archive"
                ],
                "summary": "Delete a specific archive file version",
                "operationId": "deleteArchiveFile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "fileId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/archive/{entryId}": {
            "get": {
                "tags": [
                    "Archive"
                ],
                "summary": "Get archive entry details",
                "operationId": "getArchiveId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entry ID",
                        "name": "entryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ArchiveEntryResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Archive"
                ],
                "summary": "Update archive entry (new version)",
                "operationId": "putArchiveId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entry ID",
                        "name": "entryId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateArchiveEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ArchiveEntryResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Archive"
                ],
                "summary": "Delete archive entry",
                "operationId": "deleteArchiveId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entry ID",
                        "name": "entryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/archive/{entryId}/file": {
            "get": {
                "tags": [
                    "Archive"
                ],
                "summary": "Download archive file",
                "operationId": "getArchiveFile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entry ID",
                        "name": "entryId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Specific File ID (ignored in simple mode)",
                        "name": "file_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/archive/{entryId}/versions": {
            "get": {
                "tags": [
                    "Archive"
                ],
                "summary": "List archive entry versions",
                "operationId": "getArchiveVersions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entry ID",
                        "name": "entryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ArchiveEntryResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ArchiveEntryResponse": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "edit_version": {
                    "type": "integer"
                },
                "exam_date": {
                    "type": "string"
                },
                "file_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "module_id": {
                    "type": "string"
                },
                "module_name": {
                    "type": "string"
                },
                "program_id": {
                    "type": "string"
                },
                "uploader_name": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateArchiveEntryRequest": {
            "type": "object",
            "required": [
                "date",
                "module_id",
                "version"
            ],
            "properties": {
                "comment": {
                    "type": "string",
                    "maxLength": 1000
                },
                "date": {
                    "type": "string"
                },
                "module_id": {
                    "type": "string"
                },
                "version": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                }
            }
        },
...
    }
}